Split out interest expression and login
