+++ src/stores/defaultData.ts
@@ -358,6 +358,14 @@ const getColdStartData = (): taskI[] => {
 const hiddenBody: string = '* Task Removed *';
 const IDLE_RECONNECT = hourInMillis;
 
+const statusMsgs = {
+  noAuth: 'Not Authorized to sync',
+  noProfile: 'No profile set (Not authorized to sync)',
+  interest: 'Interest expressed (Not yet authorized)',
+  renewal: 'Checking for renewal',
+  auth: 'Authorized to sync',
+};
+
 export {
   genesisTask,
   startingVelocity,
@@ -391,4 +399,5 @@ export {
   hiddenBody,
   getParentRange,
   IDLE_RECONNECT,
+  statusMsgs,
 };
